# sync to 1.1: parallelize `getTableStats` when `show accounts`